refactor-07.5: Converted to a single file - part 05
