+++ src/mcp_server/observability_mcp.py
@@ -46,7 +46,7 @@ class ObservabilityMCPServer:
             find_best_metric_with_metadata_v2,  # Smart metric selection v2
             find_best_metric_with_metadata,   # Smart metric selection v1
         )
-        from .tools.tempo import (
+        from .tools.tempo_tools import (
             query_tempo_tool,
             get_trace_details_tool,
             chat_tempo_tool,
