+++ tests/mcp_server/test_tempo_tools.py
@@ -6,7 +6,7 @@ import os
 from unittest.mock import patch, MagicMock
 from datetime import datetime, timedelta
 
-import src.mcp_server.tools.tempo as tempo_tools
+import src.mcp_server.tools.tempo_tools as tempo_tools
 
 
 def _texts(result):
@@ -17,7 +17,7 @@ def _texts(result):
 class TestQueryTempoTool:
     """Test query_tempo_tool function."""
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_query_tempo_tool_success(self, mock_tempo_class):
         """Test successful trace query."""
         # Mock the TempoQueryTool instance
@@ -68,7 +68,7 @@ class TestQueryTempoTool:
         assert "ui" in text
         assert "1500" in text
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_query_tempo_tool_no_traces(self, mock_tempo_class):
         """Test query with no traces found."""
         mock_tempo = MagicMock()
@@ -98,7 +98,7 @@ class TestQueryTempoTool:
         text = result[0]["text"]
         assert "No traces found" in text or "0 traces" in text
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_query_tempo_tool_error(self, mock_tempo_class):
         """Test query with error response."""
         mock_tempo = MagicMock()
@@ -130,7 +130,7 @@ class TestQueryTempoTool:
 class TestGetTraceDetailsTool:
     """Test get_trace_details_tool function."""
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_get_trace_details_success(self, mock_tempo_class):
         """Test successful trace details retrieval."""
         mock_tempo = MagicMock()
@@ -168,7 +168,7 @@ class TestGetTraceDetailsTool:
         assert "GET /api/users" in text
         assert "ui" in text
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_get_trace_details_error(self, mock_tempo_class):
         """Test trace details with error response."""
         mock_tempo = MagicMock()
@@ -194,7 +194,7 @@ class TestGetTraceDetailsTool:
 class TestChatTempoTool:
     """Test chat_tempo_tool function."""
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_chat_tempo_tool_success(self, mock_tempo_class):
         """Test successful conversational trace analysis."""
         mock_tempo = MagicMock()
@@ -226,7 +226,7 @@ class TestChatTempoTool:
         assert "abc123" in text
         assert "ui" in text
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_chat_tempo_tool_no_traces(self, mock_tempo_class):
         """Test conversational analysis with no traces."""
         mock_tempo = MagicMock()
@@ -248,7 +248,7 @@ class TestChatTempoTool:
         text = result[0]["text"]
         assert "No traces found" in text or "0 traces" in text
 
-    @patch("src.mcp_server.tools.tempo.mcp_tools.TempoQueryTool")
+    @patch("src.mcp_server.tools.tempo_tools.TempoQueryTool")
     def test_chat_tempo_tool_error(self, mock_tempo_class):
         """Test conversational analysis with error."""
         mock_tempo = MagicMock()
@@ -298,7 +298,7 @@ class TestTempoQueryToolClass:
     @patch("httpx.AsyncClient")
     def test_tempo_query_tool_initialization(self, mock_client):
         """Test TempoQueryTool initialization with centralized configuration."""
-        from src.mcp_server.tools.tempo import TempoQueryTool
+        from src.mcp_server.tools.tempo_tools import TempoQueryTool
         tool = TempoQueryTool()
         
         # The refactored code now uses centralized service
@@ -310,7 +310,7 @@ class TestTempoQueryToolClass:
     @patch("httpx.AsyncClient")
     def test_tempo_query_tool_default_config(self, mock_client):
         """Test TempoQueryTool with default configuration."""
-        from src.mcp_server.tools.tempo import TempoQueryTool
+        from src.mcp_server.tools.tempo_tools import TempoQueryTool
         tool = TempoQueryTool()
         
         # The refactored code now uses centralized service
@@ -324,7 +324,7 @@ class TestTempoQueryToolClass:
         """Test successful service discovery."""
         mock_get_available_services.return_value = ["ui", "api", "database"]
         
-        from src.mcp_server.tools.tempo import TempoQueryTool
+        from src.mcp_server.tools.tempo_tools import TempoQueryTool
         tool = TempoQueryTool()
         
         import asyncio
@@ -337,7 +337,7 @@ class TestTempoQueryToolClass:
         """Test service discovery with error response."""
         mock_get_available_services.return_value = []
         
-        from src.mcp_server.tools.tempo import TempoQueryTool
+        from src.mcp_server.tools.tempo_tools import TempoQueryTool
         tool = TempoQueryTool()
         
         import asyncio
